[hu192] Se agrega refactor entities
