+++ src/main/java/com/conexionmysql/mysqljdb/application/dto/ResponseFormat.java
@@ -1,4 +1,4 @@
-package com.conexionmysql.mysqljdb.infrastructure.controllers;
+package com.conexionmysql.mysqljdb.application.dto;
 
 import java.time.LocalDateTime;
 
@@ -7,11 +7,17 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+/*
+ * @author Alexander Nova
+ * @version 1.0
+ * @since 2022-03-22
+ * This class is used to format the response of the API
+ */
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ErrorResponse {
+public class ResponseFormat {
     private String message;
     private int status;
     private LocalDateTime timestamp;
