+++ src/main/java/com/conexionmysql/mysqljdb/application/services/reservation/SearchReservationService.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.conexionmysql.mysqljdb.application.dto.ReservationSearchDto;
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.Reservation;
+import com.conexionmysql.mysqljdb.domain.entities.Reservation;
 import com.conexionmysql.mysqljdb.infrastructure.jpa.repositories.ReservationRepository;
 
 @Service
