+++ src/main/java/com/conexionmysql/mysqljdb/application/services/room/GetRoomByIdService.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.Room;
+import com.conexionmysql.mysqljdb.domain.entities.Room;
 import com.conexionmysql.mysqljdb.infrastructure.jpa.repositories.RoomRepository;
 
 @Service
