+++ src/main/java/com/conexionmysql/mysqljdb/application/services/user/GetAllUserService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.User;
+import com.conexionmysql.mysqljdb.domain.entities.User;
 import com.conexionmysql.mysqljdb.infrastructure.jpa.repositories.UserRepository;
 
 @Service
