+++ src/main/java/com/conexionmysql/mysqljdb/application/services/user/GetUserByIdService.java
@@ -1,7 +1,7 @@
 package com.conexionmysql.mysqljdb.application.services.user;
 import org.springframework.stereotype.Service;
 
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.User;
+import com.conexionmysql.mysqljdb.domain.entities.User;
 import com.conexionmysql.mysqljdb.infrastructure.jpa.repositories.UserRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
