+++ src/main/java/com/conexionmysql/mysqljdb/application/services/user/UpdateUserService.java
@@ -3,7 +3,7 @@ package com.conexionmysql.mysqljdb.application.services.user;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.User;
+import com.conexionmysql.mysqljdb.domain.entities.User;
 import com.conexionmysql.mysqljdb.infrastructure.jpa.repositories.UserRepository;
 
 @Service
