+++ src/main/java/com/conexionmysql/mysqljdb/domain/entities/RoomType.java
@@ -1,4 +1,4 @@
-package com.conexionmysql.mysqljdb.infrastructure.jpa.entities;
+package com.conexionmysql.mysqljdb.domain.entities;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -10,16 +10,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-
 @Entity
-@Table(name = "users_state")
+@Table(name = "rooms_type")
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
-public class UserState {
+public class RoomType {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long user_state_id; 
+  private Long rooms_type_id; 
   private String state;
 }
