+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/controllers/reservation/ReservationController.java
@@ -2,10 +2,8 @@ package com.conexionmysql.mysqljdb.infrastructure.controllers.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.List;
-import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -20,14 +18,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.conexionmysql.mysqljdb.application.dto.ReservationCreateDto;
+import com.conexionmysql.mysqljdb.application.dto.ResponseFormat;
 import com.conexionmysql.mysqljdb.application.dto.ReservationSearchDto;
 import com.conexionmysql.mysqljdb.application.services.reservation.CreateReservationService;
 import com.conexionmysql.mysqljdb.application.services.reservation.SearchReservationService;
 import com.conexionmysql.mysqljdb.application.services.room.GetRoomByIdService;
 import com.conexionmysql.mysqljdb.application.services.user.GetUserByIdService;
-import com.conexionmysql.mysqljdb.infrastructure.controllers.ErrorResponse;
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.Reservation;
+import com.conexionmysql.mysqljdb.domain.entities.Reservation;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -55,51 +52,52 @@ public class ReservationController {
     // Verificar si el usuario existe, si se proporciona userId
     if (reservation.getUser_id() != null
         && getUserByIdService.getUserById(reservation.getUser_id().getUser_id()) == null) {
-      ErrorResponse errorResponse = new ErrorResponse(
+      ResponseFormat responseFormat = new ResponseFormat(
           "Oops, no se encontró el usuario " + reservation.getUser_id().getUser_id(),
           HttpStatus.NOT_FOUND.value(),
           LocalDateTime.now());
-      return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
+      return new ResponseEntity<>(responseFormat, HttpStatus.NOT_FOUND);
     }
 
     // Verificar si la habitación existe, si se proporciona roomId
     if (reservation.getRoom_id() != null
         && getRoomByIdService.getRoomById(reservation.getRoom_id().getRoom_id()).isEmpty()) {
-      ErrorResponse errorResponse = new ErrorResponse(
+      ResponseFormat responseFormat = new ResponseFormat(
           "Oops, no se encontró la habitación " + reservation.getRoom_id().getRoom_id(),
           HttpStatus.NOT_FOUND.value(),
           LocalDateTime.now());
-      return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
+      return new ResponseEntity<>(responseFormat, HttpStatus.NOT_FOUND);
     }
 
     // Verificar si las fechas son válidas
     if (reservation.getCheck_in_date() != null && reservation.getCheck_out_date() != null
         && reservation.getCheck_in_date().isAfter(reservation.getCheck_out_date())) {
-      ErrorResponse errorResponse = new ErrorResponse("La fecha de check-in debe ser anterior a la fecha de check-out",
+      ResponseFormat responseFormat = new ResponseFormat(
+          "La fecha de check-in debe ser anterior a la fecha de check-out",
           HttpStatus.BAD_REQUEST.value(),
           LocalDateTime.now());
-      return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
+      return new ResponseEntity<>(responseFormat, HttpStatus.BAD_REQUEST);
     }
 
     try {
       createReservationService.createReservation(reservation);
-      ErrorResponse Response = new ErrorResponse("Se creoo bien",
+      ResponseFormat Response = new ResponseFormat("Se creoo bien",
           HttpStatus.IM_USED.value(),
           LocalDateTime.now());
       return new ResponseEntity<>(Response, HttpStatus.CREATED);
     } catch (DateTimeParseException e) {
 
-      ErrorResponse errorResponse = new ErrorResponse("Invalid date format. Date format should be yyyy-MM-dd HH:mm:ss",
+      ResponseFormat responseFormat = new ResponseFormat(
+          "Invalid date format. Date format should be yyyy-MM-dd HH:mm:ss",
           HttpStatus.BAD_REQUEST.value(), LocalDateTime.now());
-      return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
+      return new ResponseEntity<>(responseFormat, HttpStatus.BAD_REQUEST);
     } catch (DataIntegrityViolationException e) {
-      ErrorResponse errorResponse = new ErrorResponse("Error creating reservation: ",
+      ResponseFormat responseFormat = new ResponseFormat("Error creating reservation: ",
           HttpStatus.BAD_REQUEST.value(), LocalDateTime.now());
-      return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
+      return new ResponseEntity<>(responseFormat, HttpStatus.BAD_REQUEST);
     }
   }
 
-  
   @GetMapping
   public ResponseEntity<Object> searchReservations(
       @RequestParam(value = "user_id", required = false) Long userId,
@@ -112,15 +110,15 @@ public class ReservationController {
 
     // Check if user exists
     if (userId != null && getUserByIdService.getUserById(userId) == null) {
-      ErrorResponse errorResponse = new ErrorResponse("Oops, no se encontró el usuario " + userId,
+      ResponseFormat responseFormat = new ResponseFormat("Oops, no se encontró el usuario " + userId,
           HttpStatus.NOT_FOUND.value(),
           LocalDateTime.now());
-      return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
+      return new ResponseEntity<>(responseFormat, HttpStatus.NOT_FOUND);
     }
 
     // Check if room exists
     if (roomId != null && getRoomByIdService.getRoomById(roomId).isEmpty()) {
-      ErrorResponse errorResponse = new ErrorResponse("Oops, no se encontró la habitación " + roomId,
+      ResponseFormat errorResponse = new ResponseFormat("Oops, no se encontró la habitación " + roomId,
           HttpStatus.NOT_FOUND.value(),
           LocalDateTime.now());
       return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
@@ -134,7 +132,7 @@ public class ReservationController {
       try {
         checkInDate = LocalDate.parse(checkInDateString).atStartOfDay();
       } catch (DateTimeParseException e) {
-        ErrorResponse errorResponse = new ErrorResponse(
+        ResponseFormat errorResponse = new ResponseFormat(
             "Invalid check-in date format. Date format should be yyyy-MM-dd",
             HttpStatus.BAD_REQUEST.value(), LocalDateTime.now());
         return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
@@ -146,7 +144,7 @@ public class ReservationController {
       try {
         checkOutDate = LocalDate.parse(checkOutDateString).atStartOfDay();
       } catch (DateTimeParseException e) {
-        ErrorResponse errorResponse = new ErrorResponse(
+        ResponseFormat errorResponse = new ResponseFormat(
             "Invalid check-out date format. Date format should be yyyy-MM-dd",
             HttpStatus.BAD_REQUEST.value(), LocalDateTime.now());
         return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
@@ -155,7 +153,8 @@ public class ReservationController {
 
     // Check if dates are valid
     if (checkInDate != null && checkOutDate != null && checkInDate.isAfter(checkOutDate)) {
-      ErrorResponse errorResponse = new ErrorResponse("La fecha de check-in debe ser anterior a la fecha de check-out",
+      ResponseFormat errorResponse = new ResponseFormat(
+          "La fecha de check-in debe ser anterior a la fecha de check-out",
           HttpStatus.BAD_REQUEST.value(),
           LocalDateTime.now());
       return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
@@ -172,7 +171,7 @@ public class ReservationController {
       List<Reservation> reservations = searchReservationService.searchReservations(reservationSearchDto);
       return new ResponseEntity<>(reservations, HttpStatus.OK);
     } catch (Exception e) {
-      ErrorResponse errorResponse = new ErrorResponse("Error searching reservations",
+      ResponseFormat errorResponse = new ResponseFormat("Error searching reservations",
           HttpStatus.INTERNAL_SERVER_ERROR.value(),
           LocalDateTime.now());
       return new ResponseEntity<>(errorResponse, HttpStatus.INTERNAL_SERVER_ERROR);
