+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/controllers/room/RoomController.java
@@ -1,0 +1,5 @@
+package com.conexionmysql.mysqljdb.infrastructure.controllers.room;
+
+public class RoomController {
+  
+}
