+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/jpa/repositories/ReservationStateRepository.java
@@ -2,7 +2,7 @@ package com.conexionmysql.mysqljdb.infrastructure.jpa.repositories;
 
 import org.springframework.stereotype.Repository;
 
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.ReservationState;
+import com.conexionmysql.mysqljdb.domain.entities.ReservationState;
 
 import org.springframework.data.jpa.repository.JpaRepository;
