+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/jpa/repositories/RoomRepository.java
@@ -3,7 +3,7 @@ package com.conexionmysql.mysqljdb.infrastructure.jpa.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.Room;
+import com.conexionmysql.mysqljdb.domain.entities.Room;
 
 @Repository
 public interface RoomRepository extends JpaRepository<Room, Long> {
