+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/jpa/repositories/RoomStateRepository.java
@@ -3,7 +3,7 @@ package com.conexionmysql.mysqljdb.infrastructure.jpa.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.RoomState;
+import com.conexionmysql.mysqljdb.domain.entities.RoomState;
 
 @Repository
 public interface RoomStateRepository extends JpaRepository<RoomState, Long> {
