+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/jpa/repositories/RoomTypeRepository.java
@@ -3,7 +3,7 @@ package com.conexionmysql.mysqljdb.infrastructure.jpa.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.RoomType;
+import com.conexionmysql.mysqljdb.domain.entities.RoomType;
 
 @Repository
 public interface RoomTypeRepository extends JpaRepository<RoomType, Long> {
