+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/jpa/repositories/UserRepository.java
@@ -3,7 +3,7 @@ package com.conexionmysql.mysqljdb.infrastructure.jpa.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.User;
+import com.conexionmysql.mysqljdb.domain.entities.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
