+++ src/main/java/com/conexionmysql/mysqljdb/infrastructure/jpa/repositories/UserStateRepository.java
@@ -3,7 +3,7 @@ package com.conexionmysql.mysqljdb.infrastructure.jpa.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.conexionmysql.mysqljdb.infrastructure.jpa.entities.UserState;
+import com.conexionmysql.mysqljdb.domain.entities.UserState;
 
 @Repository
 public interface UserStateRepository extends JpaRepository<UserState, Long> {
